Move IkFast header to dart/dynamics
This commit moves ikfast.h from dart/external/ikfast/ to dart/dynamics/
as part of the effort to remove the dart/external/ directory.

Changes:
- Move dart/external/ikfast/ikfast.h → dart/dynamics/ikfast.h
- Update includes in DART core (IkFast.hpp, IkFast.cpp)
- Update includes in examples and tests
- Remove dart/external/ikfast/ directory
- Add ikfast.h to install in dart/dynamics/CMakeLists.txt

Benefits:
- Better organization: IkFast header lives with implementation
- Cleaner install: Header installed to include/dart/dynamics/
- No external path exposure in public API
