+++ dart/dynamics/IkFast.cpp
@@ -35,7 +35,7 @@
 #include "dart/dynamics/BodyNode.hpp"
 #include "dart/dynamics/DegreeOfFreedom.hpp"
 #include "dart/dynamics/RevoluteJoint.hpp"
-#include "dart/external/ikfast/ikfast.h"
+#include "dart/dynamics/ikfast.h"
 
 namespace dart {
 namespace dynamics {
