+++ dart/external/CMakeLists.txt
@@ -2,5 +2,4 @@ add_subdirectory(convhull_3d)
 if(NOT DART_USE_SYSTEM_IMGUI)
   add_subdirectory(imgui)
 endif()
-add_subdirectory(ikfast)
 add_subdirectory(odelcpsolver)
