+++ tests/integration/GeneratedWamIkFast.cpp
@@ -20,7 +20,7 @@
 ///     gcc -fPIC -lstdc++ -DIKFAST_NO_MAIN -DIKFAST_CLIBRARY -shared
 ///     -Wl,-soname,libik.so -o libik.so ik.cpp
 #define IKFAST_HAS_LIBRARY
-#include "dart/external/ikfast/ikfast.h" // found inside share/openrave-X.Y/python/ikfast.h
+#include <dart/dynamics/ikfast.h> // found inside share/openrave-X.Y/python/ikfast.h
 using namespace ikfast;
 
 // check if the included ikfast version matches what this file was compiled with
